// src/types/IUser.ts
export interface IuserData {
  touchIDPreference: string;
  appState: string;
  email: string;
  aggregateVerifier: string;
  name: string;
  profileImage: string;
  typeOfLogin: string;
  verifier: string;
  verifierId: string;
  dappShare: string;
  oAuthIdToken: string;
  oAuthAccessToken: string;
  idToken: string;
  wallet: string;
  balance: string;
}